Add function that retrieves all entities (edge/face) of the body
The dataset represents labels (joints) as relation of the entities
and we represent entites as EntityOne structures.
Labels were calculated with Fusion360 cad tool, we use OpenCascade here,
so to find the match between dataset label's entity and our body
entities would be convinent to have simialar representaiton fo the data.